tests(app): 💸 outline `app_can_spend_notes_and_detect_outputs`
(NB: renamed from `mock_consensus_can_spend_notes_and_detect_outputs`)
